Use doubles for particles until they make it to the GPU
This actually results in a very large performance update (using release optimizations) since time requires double precision to be accurate, and every operation involving time and particles would require a conversion from floats to double.

+++ src/entities/particle.h
@@ -3,7 +3,7 @@
 namespace rpg {
 
 struct Particle {
-  using Vector3 = std::array<float, 3>;
+  using Vector3 = std::array<double, 3>;
   float lifetime;
   Vector3 pos;
   Vector3 color;
